requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Google Cloud and Vertex AI
google-cloud-aiplatform>=1.38.0
google-auth>=2.27.0
vertexai>=1.38.0

# Vector database
faiss-cpu>=1.7.4
numpy>=1.24.0,<2.0.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Data validation and models
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
structlog>=24.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0

# Evaluation and metrics
pandas>=2.1.0
scikit-learn>=1.3.0

# Type hints
typing-extensions>=4.9.0
